move numbers<> to its own header
