+++ include/metal/number.hpp
@@ -22,6 +22,7 @@
 #include <metal/number/neg.hpp>
 #include <metal/number/not.hpp>
 #include <metal/number/number.hpp>
+#include <metal/number/numbers.hpp>
 #include <metal/number/or.hpp>
 #include <metal/number/pow.hpp>
 #include <metal/number/sub.hpp>
