+++ include/metal/number/min.hpp
@@ -46,6 +46,7 @@ namespace metal
 }
 
 #include <metal/number/number.hpp>
+#include <metal/number/numbers.hpp>
 #include <metal/lambda/lambda.hpp>
 #include <metal/list/accumulate.hpp>
 #include <metal/list/list.hpp>
