+++ include/metal/number/number.hpp
@@ -19,9 +19,6 @@ namespace metal
         struct _is_number;
 
         using int_ = std::intmax_t;
-
-        template<int_... vs>
-        struct _numbers;
     }
     /// \endcond
 
@@ -99,29 +96,6 @@ namespace metal
     /// ### See Also
     /// \see number
     using false_ = metal::number<false>;
-}
-
-#include <metal/list/list.hpp>
-
-namespace metal
-{
-    /// \ingroup number
-    ///
-    /// ### Description
-    /// Constructs a \list of \numbers out of a sequence of integral values.
-    ///
-    /// ### Example
-    /// \snippet number.cpp numbers
-    ///
-    /// ### See Also
-    /// \see int_, number, list
-    template<int_... vs>
-    using numbers =
-#if defined(METAL_COMPAT_MODE) && !defined(METAL_DOXYGENATING)
-        typename detail::_numbers<vs...>::type;
-#else
-        metal::list<metal::number<vs>...>;
-#endif
 
     /// \cond
     namespace detail
@@ -135,14 +109,6 @@ namespace metal
         struct _is_number<number<value>> :
             true_
         {};
-
-#if defined(METAL_COMPAT_MODE)
-        template<int_... vs>
-        struct _numbers
-        {
-            using type = list<std::integral_constant<int_, vs>...>;
-        };
-#endif
     }
     /// \endcond
 }
