+++ include/metal/number/pow.hpp
@@ -52,6 +52,7 @@ namespace metal
 }
 
 #include <metal/number/number.hpp>
+#include <metal/number/numbers.hpp>
 #include <metal/lambda/lambda.hpp>
 #include <metal/list/list.hpp>
 #include <metal/list/accumulate.hpp>
